Move core init into shared module and ignore build artifacts
