--- src/audio_capture.d
@@ -1,2 +1,0 @@
-src/audio_capture.o: src/audio_capture.c src/../include/rootstream.h
-src/../include/rootstream.h:
