--- src/audio_playback.d
@@ -1,2 +1,0 @@
-src/audio_playback.o: src/audio_playback.c src/../include/rootstream.h
-src/../include/rootstream.h:
