--- src/config.d
@@ -1,2 +1,0 @@
-src/config.o: src/config.c src/../include/rootstream.h
-src/../include/rootstream.h:
