--- src/crypto.d
@@ -1,2 +1,0 @@
-src/crypto.o: src/crypto.c src/../include/rootstream.h
-src/../include/rootstream.h:
