--- src/discovery.d
@@ -1,2 +1,0 @@
-src/discovery.o: src/discovery.c src/../include/rootstream.h
-src/../include/rootstream.h:
