--- src/display_sdl2.d
@@ -1,2 +1,0 @@
-src/display_sdl2.o: src/display_sdl2.c src/../include/rootstream.h
-src/../include/rootstream.h:
