--- src/drm_capture.d
@@ -1,2 +1,0 @@
-src/drm_capture.o: src/drm_capture.c src/../include/rootstream.h
-src/../include/rootstream.h:
