--- src/input.d
@@ -1,2 +1,0 @@
-src/input.o: src/input.c src/../include/rootstream.h
-src/../include/rootstream.h:
