--- src/latency.d
@@ -1,2 +1,0 @@
-src/latency.o: src/latency.c src/../include/rootstream.h
-src/../include/rootstream.h:
