--- src/main.d
@@ -1,2 +1,0 @@
-src/main.o: src/main.c src/../include/rootstream.h
-src/../include/rootstream.h:
