--- src/network.d
@@ -1,2 +1,0 @@
-src/network.o: src/network.c src/../include/rootstream.h
-src/../include/rootstream.h:
