--- src/nvenc_encoder.d
@@ -1,2 +1,0 @@
-src/nvenc_encoder.o: src/nvenc_encoder.c src/../include/rootstream.h
-src/../include/rootstream.h:
