--- src/opus_codec.d
@@ -1,2 +1,0 @@
-src/opus_codec.o: src/opus_codec.c src/../include/rootstream.h
-src/../include/rootstream.h:
