--- src/qrcode.d
@@ -1,2 +1,0 @@
-src/qrcode.o: src/qrcode.c src/../include/rootstream.h
-src/../include/rootstream.h:
