--- src/service.d
@@ -1,2 +1,0 @@
-src/service.o: src/service.c src/../include/rootstream.h
-src/../include/rootstream.h:
