--- src/tray_stub.d
@@ -1,2 +1,0 @@
-src/tray_stub.o: src/tray_stub.c src/../include/rootstream.h
-src/../include/rootstream.h:
