--- src/vaapi_decoder.d
@@ -1,2 +1,0 @@
-src/vaapi_decoder.o: src/vaapi_decoder.c src/../include/rootstream.h
-src/../include/rootstream.h:
