--- src/vaapi_encoder.d
@@ -1,2 +1,0 @@
-src/vaapi_encoder.o: src/vaapi_encoder.c src/../include/rootstream.h
-src/../include/rootstream.h:
